+++ flink/flink-1.14/src/test/java/io/tidb/bigdata/flink/tidb/stream/delete/TiKVDeleteTest.java
@@ -13,7 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.tidb.bigdata.flink.tidb.delete;
+
+package io.tidb.bigdata.flink.tidb.stream.delete;
 
 import static io.tidb.bigdata.flink.connector.TiDBOptions.DELETE_ENABLE;
 import static io.tidb.bigdata.flink.connector.TiDBOptions.IGNORE_PARSE_ERRORS;
@@ -21,8 +22,7 @@ import static io.tidb.bigdata.flink.connector.TiDBOptions.SINK_BUFFER_SIZE;
 import static io.tidb.bigdata.flink.connector.TiDBOptions.SINK_IMPL;
 import static io.tidb.bigdata.flink.connector.TiDBOptions.SINK_TRANSACTION;
 import static io.tidb.bigdata.flink.connector.TiDBOptions.STREAMING_CODEC;
-import static io.tidb.bigdata.flink.connector.TiDBOptions.STREAMING_CODEC_CANAL_JSON;
-import static io.tidb.bigdata.flink.connector.TiDBOptions.STREAMING_CODEC_CRAFT;
+import static io.tidb.bigdata.flink.connector.TiDBOptions.STREAMING_CODEC_JSON;
 import static io.tidb.bigdata.flink.connector.TiDBOptions.STREAMING_SOURCE;
 import static io.tidb.bigdata.flink.connector.TiDBOptions.SinkImpl.TIKV;
 import static io.tidb.bigdata.flink.connector.TiDBOptions.WRITE_MODE;
@@ -53,46 +53,56 @@ import org.junit.experimental.categories.Category;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized.Parameters;
 
-/** test for different codec, json has been test in {@link TiKVDeleteTest} */
+/** Test for pk, uk and multiple_uk table */
 @Category(IntegrationTest.class)
 @RunWith(org.junit.runners.Parameterized.class)
-public class TiKVDeleteCodecTest extends FlinkTestBase {
+public class TiKVDeleteTest extends FlinkTestBase {
+
   private String srcTable;
 
   private String dstTable;
 
-  public TiKVDeleteCodecTest(
-      String flinkDeleteTable, int result, String kafkaGroup, String topic, String streamingCodec) {
+  public TiKVDeleteTest(String flinkDeleteTable, int result, String kafkaGroup) {
     this.flinkDeleteTable = flinkDeleteTable;
     this.result = result;
     this.kafkaGroup = kafkaGroup;
-    this.topic = topic;
-    this.streamingCodec = streamingCodec;
   }
 
-  @Parameters(
-      name =
-          "{index}: FlinkDeleteTable={0}, Result={1} ,KafkaGroup={2},Topic={3},StreamingCodec={4}")
+  @Parameters(name = "{index}: FlinkDeleteTable={0}, Result={1} ,KafkaGroup={2} ")
   public static Collection<Object[]> data() {
     return Arrays.asList(
         new Object[][] {
-          {TABLE_PK, 2, "group_c_1", "tidb_test_craft", STREAMING_CODEC_CRAFT},
-          {TABLE_PK, 2, "group_c_2", "tidb_test_canal_json", STREAMING_CODEC_CANAL_JSON},
+          {TABLE_PK, 2, "group_d_1"}, {TABLE_UK, 2, "group_d_2"}, {TABLE_MUTILE_UK, 2, "group_d_3"},
         });
   }
 
   private final String flinkDeleteTable;
   private final int result;
   private final String kafkaGroup;
-  private final String topic;
-  private final String streamingCodec;
 
   private static final String TABLE_PK =
       "CREATE TABLE IF NOT EXISTS `%s`.`%s`\n"
           + "(\n"
-          + "    c1  bigint(20),\n"
+          + "    c1  varchar(255),\n"
           + "    c2  varchar(255),\n"
           + "    PRIMARY KEY (`c1`) \n"
+          + ")";
+
+  private static final String TABLE_UK =
+      "CREATE TABLE IF NOT EXISTS `%s`.`%s`\n"
+          + "(\n"
+          + "    c1  bigint(20) NULL DEFAULT NULL,\n"
+          + "    c2  bigint(20) NOT NULL,\n"
+          + "    UNIQUE INDEX `uniq_1`(`c1`) USING BTREE,\n"
+          + "    UNIQUE INDEX `uniq_2`(`c2`) USING BTREE\n"
+          + ")";
+
+  private static final String TABLE_MUTILE_UK =
+      "CREATE TABLE IF NOT EXISTS `%s`.`%s`\n"
+          + "(\n"
+          + "    c1  bigint(20) NOT NULL,\n"
+          + "    c2  bigint(20) NOT NULL,\n"
+          + "    UNIQUE INDEX `uniq_1`(`c1`,`c2`) USING BTREE\n"
           + ")";
 
   @Test
@@ -105,9 +115,9 @@ public class TiKVDeleteCodecTest extends FlinkTestBase {
     Map<String, String> properties = defaultProperties();
     // source
     properties.put(STREAMING_SOURCE.key(), "kafka");
-    properties.put(STREAMING_CODEC.key(), streamingCodec);
+    properties.put(STREAMING_CODEC.key(), STREAMING_CODEC_JSON);
     properties.put("tidb.streaming.kafka.bootstrap.servers", "localhost:9092");
-    properties.put("tidb.streaming.kafka.topic", topic);
+    properties.put("tidb.streaming.kafka.topic", "tidb_test");
     properties.put("tidb.streaming.kafka.group.id", kafkaGroup);
     properties.put(IGNORE_PARSE_ERRORS.key(), "true");
     // sink
@@ -146,6 +156,7 @@ public class TiKVDeleteCodecTest extends FlinkTestBase {
     Thread.sleep(20000);
 
     JobClient jobClient = task.get().orElseThrow(IllegalStateException::new);
+
     // insert 4 rows in src, flush 3 insert to dst
     testDatabase
         .getClientSession()
@@ -158,7 +169,6 @@ public class TiKVDeleteCodecTest extends FlinkTestBase {
         4, testDatabase.getClientSession().queryTableCount(DATABASE_NAME, srcTable));
     Assert.assertEquals(
         3, testDatabase.getClientSession().queryTableCount(DATABASE_NAME, dstTable));
-
     // delete 3 rows in src , flush 1 insert and 2 delete to dst
     testDatabase
         .getClientSession()
